+++ src/duties/lint.spec.js
@@ -3,14 +3,22 @@
 let lint = require('./lint')
 let lintHelpers = require('../test-data/lint-results')
 
+let danger = {
+  github: {
+    pr: {
+      head: { ref: 'master', repo: { html_url: 'https://test.example' } },
+    },
+  },
+}
+
 describe('lint results', () => {
   it('allows lint results to be undefined', () => {
     let fail = jest.fn()
     let warn = jest.fn()
     let message = jest.fn()
     let lintResults = undefined
-    lint.hasLintErrors({ lintResults, fail, message })
-    lint.hasLintWarnings({ lintResults, warn, message })
+    lint.hasLintErrors({ lintResults, fail, message, danger })
+    lint.hasLintWarnings({ lintResults, warn, message, danger })
 
     expect(fail).not.toHaveBeenCalled()
     expect(warn).not.toHaveBeenCalled()
@@ -20,7 +28,7 @@ describe('lint results', () => {
   it('fails if they have errors', () => {
     let fail = jest.fn()
     let lintResults = lintHelpers.failing
-    lint.hasLintErrors({ lintResults, fail })
+    lint.hasLintErrors({ lintResults, fail, danger })
 
     expect(fail).toHaveBeenCalled()
   })
@@ -28,7 +36,7 @@ describe('lint results', () => {
   it('warns if they have warnings', () => {
     let warn = jest.fn()
     let lintResults = lintHelpers.warning
-    lint.hasLintWarnings({ lintResults, warn })
+    lint.hasLintWarnings({ lintResults, warn, danger })
 
     expect(warn).toHaveBeenCalled()
   })
@@ -37,8 +45,8 @@ describe('lint results', () => {
     let warn = jest.fn()
     let fail = jest.fn()
     let lintResults = lintHelpers.passing
-    lint.hasLintWarnings({ lintResults, warn })
-    lint.hasLintErrors({ lintResults, fail })
+    lint.hasLintWarnings({ lintResults, warn, danger })
+    lint.hasLintErrors({ lintResults, fail, danger })
 
     expect(warn).not.toHaveBeenCalled()
     expect(fail).not.toHaveBeenCalled()
@@ -48,7 +56,7 @@ describe('lint results', () => {
     it('passes if there arent any messages', () => {
       let fail = jest.fn()
       let lintResults = lintHelpers.standardJsonPassing
-      lint.hasLintErrors({ lintResults, fail })
+      lint.hasLintErrors({ lintResults, fail, danger })
 
       expect(fail).not.toHaveBeenCalled()
     })
@@ -56,9 +64,33 @@ describe('lint results', () => {
     it('fails if there are any messages', () => {
       let fail = jest.fn()
       let lintResults = lintHelpers.standardJsonFailing
-      lint.hasLintErrors({ lintResults, fail })
+      lint.hasLintErrors({ lintResults, fail, danger })
 
       expect(fail).toHaveBeenCalled()
     })
+  })
+
+  it('adds related links to lint results with warnings', () => {
+    let warn = jest.fn()
+    let lintResults = lintHelpers.warning
+
+    lint.hasLintWarnings({ lintResults, warn, danger })
+    let expected = /https:\/\/test.example\/blob\/master/.test(
+      warn.mock.calls[0][0]
+    )
+
+    expect(expected).toBe(true)
+  })
+
+  it('adds related links to lint results with errors', () => {
+    let fail = jest.fn()
+    let lintResults = lintHelpers.failing
+
+    lint.hasLintErrors({ lintResults, fail, danger })
+    let expected = /https:\/\/test.example\/blob\/master/.test(
+      fail.mock.calls[0][0]
+    )
+
+    expect(expected).toBe(true)
   })
 })