+++ src/duties/testResults.js
@@ -1,13 +1,15 @@
 let _ = require('lodash/fp')
 let stripAnsi = require('strip-ansi')
-let { basename } = require('path')
+let { pathTail, linkifyPath } = require('../utils')
 
-let resultTemplate = isMocha => r => `<details>
-  <summary>${isMocha ? r.title : basename(r.name)}</summary>
+let resultTemplate = (isMocha, danger, config) => r => `<details>
+  <summary>${isMocha
+    ? r.title
+    : linkifyPath({ path: pathTail(r.name, config), danger })}</summary>
   <code>${isMocha ? r.err.message : stripAnsi(r.message)}</code>
 </details>`
 
-let hasTestErrors = ({ testResults, fail }) => {
+let hasTestErrors = ({ testResults, fail, danger, config }) => {
   let isMocha = _.get('stats', testResults)
   if (
     testResults &&
@@ -15,7 +17,7 @@ let hasTestErrors = ({ testResults, fail }) => {
   ) {
     _.flow(
       _.reject(result => _.isEmpty(isMocha ? result.err : result.message)),
-      _.map(resultTemplate(isMocha)),
+      _.map(resultTemplate(isMocha, danger, config)),
       _.join('\n'),
       allFailures =>
         `This PR has failing tests. Please alleviate the errors and commit them\n\n${allFailures}`,